requirements.txt
# 原有依赖
torch>=1.10.2
numpy>=1.19.2
scipy>=1.5.2
scikit-learn>=0.24.0
tqdm>=4.60.0
pytorch-metric-learning==2.1.1
transformers>=4.18.0
tokenizers==0.12.1

# 新增CLIP相关依赖
clip-by-openai>=1.0
ftfy>=6.0.3
regex==2023.8.8
Pillow>=8.3.0

# 其他依赖
pytorch-metric-learning>=1.3.0
matplotlib>=3.3.4
seaborn>=0.11.0
torchvision==0.8.2
# cudatoolkit~=11.0

pandas>=1.1.5

# 可选的加速依赖
# nvidia-ml-py3  # GPU监控
# psutil  # 系统资源监控
